third_party/blink/web_tests/external/wpt/css/css-transitions/transition-behavior-expected.txt

This is a testharness.js-based test.
[FAIL] transition-behavior:normal overlaps the previous allow-discrete value.
  assert_equals: Should not start a discrete property transition expected 0 but got 1
Harness: the test ran to completion.
